requirements.txt
boto3==1.28.34
python-dotenv==1.0.0